rename player array and manager
